+++ picasso/spinna.py
@@ -508,30 +508,6 @@ def plot_NN(
         of strings is given, several paths can be specified (with
         different extensions).
     """
-
-    def remove_patches_and_data(ax, xmax):
-        # remove the bins above the xlim if given (downstream they
-        # will be displayed by PyQt somehow)
-        for patch in ax.patches:
-            left = patch.get_x()
-            right = left + patch.get_width()
-            if right > xmax:
-                #     new_patches.append(patch)
-                # else:
-                patch.remove()
-        # same for lines
-        for line in ax.lines:
-            xdata = line.get_xdata()
-            ydata = line.get_ydata()
-            new_xdata = []
-            new_ydata = []
-            for x, y in zip(xdata, ydata):
-                if x <= xmax:
-                    new_xdata.append(x)
-                    new_ydata.append(y)
-            line.set_xdata(new_xdata)
-            line.set_ydata(new_ydata)
-
     # initiate figure and axis
     if fig is None or ax is None:
         fig, ax = plt.subplots(
@@ -582,8 +558,6 @@ def plot_NN(
                 alpha=alpha,
                 label=f"sim {i+1}th NN",
             )
-    if xlim is not None:
-        remove_patches_and_data(ax, xlim[1])
 
     # display parameters
     if show_legend: