+++ changelog.rst
@@ -14,6 +14,7 @@ Important updates:
 *Small improvements:*
 +++++++++++++++++++++
 - Picasso applies constrained layout to all matplotlib figures
+- SPINNA uses ``FigureCanvas`` instead of ``QSvgRenderer`` for displaying NND plots and mask legend
 
 *Bug fixes:*
 ++++++++++++